moved `message_data` from `data/` to `tools/` folder. Changed architecture of Decoder to make data elements private instead of public to increase safety. Also added static assert to ensure setting `DecodeBufferSize` is power of 2

+++ settings.hpp
@@ -13,7 +13,7 @@ namespace ctrader::settings {
     constexpr u8           HearthBeatIntervalSec = 30;
     constexpr u8           MsgSeqNumDigitSize = 18; // range {1 ... 18}
     constexpr u8           KeySize = 6;
-    constexpr u8           DecodeBufferSize = 32; 
+    constexpr u16          DecodeBufferSize = 32; 
    
     #define __SETTINGS_SOH_STR "|"
     #define __SETTINGS_SOH '|'
